+++ contracts-periphery/src/interface/IUmbra.sol
@@ -8,13 +8,13 @@ interface IUmbra {
    * @notice Public state variable get function
    * @return uint256 toll
    */
-  function toll() external returns (uint);
+  function toll() external returns (uint256);
 
   /**
    * @notice Admin only function to update the toll
    * @param _newToll New ETH toll in wei
    */
-  function setToll(uint _newToll) external;
+  function setToll(uint256 _newToll) external;
 
   /**
    * @notice Admin only function to update the toll collector
@@ -48,7 +48,7 @@ interface IUmbra {
    */
   function sendEth(
     address payable _receiver,
-    uint _tollCommitment,
+    uint256 _tollCommitment,
     bytes32 _pkx, // ephemeral public key x coordinate
     bytes32 _ciphertext
   ) external payable;
@@ -64,7 +64,7 @@ interface IUmbra {
   function sendToken(
     address _receiver,
     address _tokenAddr,
-    uint _amount,
+    uint256 _amount,
     bytes32 _pkx, // ephemeral public key x coordinate
     bytes32 _ciphertext
   ) external payable;
@@ -112,7 +112,7 @@ interface IUmbra {
     address _acceptor,
     address _tokenAddr,
     address _sponsor,
-    uint _sponsorFee,
+    uint256 _sponsorFee,
     uint8 _v,
     bytes32 _r,
     bytes32 _s
@@ -136,7 +136,7 @@ interface IUmbra {
     address _acceptor,
     address _tokenAddr,
     address _sponsor,
-    uint _sponsorFee,
+    uint256 _sponsorFee,
     IUmbraHookReceiver _hook,
     bytes memory _data,
     uint8 _v,
